+++ v3/pb.go
@@ -593,24 +593,3 @@ func (s *State) IsFirst() bool {
 func (s *State) Time() time.Time {
 	return s.time
 }
-
-type Progressable interface {
-	Total() int64
-	Value() int64
-	Finished() bool
-}
-
-func RegisterProgressable(pr Progressable, removeFunc func(*ProgressBar)) *ProgressBar {
-	pb := new(ProgressBar)
-	go progressWorker(pr, pb, removeFunc)
-	return pb
-}
-
-func progressWorker(pr Progressable, pb *ProgressBar, removeFunc func(*ProgressBar)) {
-	for ; !pr.Finished(); time.Sleep(time.Second) {
-		_ = pb.SetTotal(pr.Total())
-		_ = pb.SetCurrent(pr.Value())
-	}
-	removeFunc(pb)
-	pb.Finish()
-}
